libfmt doesn't like our enums anymore.

+++ include/ObjectManager.h
@@ -113,18 +113,18 @@ public:
 
         auto object = ObjectBase::createObject(itemID, nextFreeObjectID, initializer);
         if (!object) {
-            sdl2::log_error("createObjectFromItemId() could not build item type {}", itemID);
+            sdl2::log_error("createObjectFromItemId() could not build item type {}", static_cast<int>(itemID));
             return nullptr;
         }
 
         auto* const pObject = dune_cast<ObjectType>(object.get());
         if (!pObject) {
-            sdl2::log_error("createObjectFromItemId() created the wrong type of object for build item type {}", itemID);
+            sdl2::log_error("createObjectFromItemId() created the wrong type of object for build item type {}", static_cast<int>(itemID));
             return nullptr;
         }
 
         if (!addObject(std::move(object))) {
-            sdl2::log_error("createObjectFromItemId() unable to add object of item type {}", itemID);
+            sdl2::log_error("createObjectFromItemId() unable to add object of item type {}", static_cast<int>(itemID));
             return nullptr;
         }
